Move RcMap to lab._rcmap #10
