+++ lab/_app.py
@@ -12,6 +12,7 @@ import numpy as np
 
 from ._bootstrap import get_current_user, open_resource, save_inputCells
 from ._plot import draw
+from ._rcmap import RcMap
 from .base import HasSource
 from .db import _schema
 from .ui import ApplicationUI, display_source_code
@@ -396,45 +397,6 @@ class SweepSet:
             sweep.parent = self
             self._sweep[sweep.name] = sweep
         return self.app
-
-
-class RcMap:
-    def __init__(self, rc={}, parent=None):
-        self.rc = {}
-        self.parent = parent
-        self.rc.update(rc)
-
-    def update(self, rc={}):
-        self.rc.update(rc)
-
-    def items(self):
-        return [(name, self.__getitem__(name)) for name in self.keys()]
-
-    def keys(self):
-        keys = set(self.rc.keys())
-        if self.parent is not None:
-            keys = keys.union(self.parent.keys())
-        return list(keys)
-
-    def get(self, name, default=None):
-        if name in self.keys():
-            return self.get_resource(name)
-        elif default is None:
-            raise KeyError('key %r not found in RcMap.' % name)
-        else:
-            return default
-
-    def get_resource(self, name):
-        name = self.rc.get(name, name)
-        if not isinstance(name, str):
-            return name
-        elif self.parent is not None:
-            return self.parent.get_resource(name)
-        else:
-            return open_resource(name)
-
-    def __getitem__(self, name):
-        return self.get(name)
 
 
 def getAppClass(name='', package='', version=None, id=None, **kwds):
